fix: compile workflow offering resolution
